Safari-609.1.11.1/LayoutTests/js/kde/evil-n-expected.txt

KDE JS Test

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS (new Error()).message is ''
PASS ''.split(/.*/).length is 0
PASS successfullyParsed is true

TEST COMPLETE
